requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
mysqlclient==2.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.1.0
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.0
SQLAlchemy==2.0.23
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
alembic==1.12.1
numpy==1.26.1
openpyxl==3.1.2
MarkupSafe==2.1.3
WTForms==3.1.1
gunicorn==21.2.0 